requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-celery==3.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
haversine==2.5.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==3.0.37
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
